+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [HexaPDF::Layout::TextFragment#text] and [HexaPDF::Layout::TextBox#text] for
   retrieving the text represented by the stored items
 * [HexaPDF::Content::Canvas#pos] for retrieving untransformed positions
+* [HexaPDF::Type::CIDFont::CIDSystemInfo] type class
 
 ### Changed